feat: 댕톡 데이터 바인딩
