+++ src/components/Social/TalkTab/styles.ts
